refactor app model
