+++ lib/components/comment_fieldbox.dart
@@ -1,7 +1,6 @@
 import 'package:brainworld/constants/constant.dart';
-import 'package:brainworld/models/user.dart';
+import 'package:brainworld/models/user_model.dart';
 import 'package:brainworld/pages/chats/components/chat_icon_gradient.dart';
-import 'package:brainworld/pages/chats/models/posts_model.dart';
 import 'package:flutter/material.dart';
 import 'package:iconly/iconly.dart';
 import 'package:provider/provider.dart';
