+++ lib/constants/constant.dart
@@ -3,9 +3,7 @@ import 'package:brainworld/components/utilities_widgets/gradient_text.dart';
 import 'package:brainworld/components/utilities_widgets/my_navigate.dart';
 import 'package:brainworld/components/utilities_widgets/radial_gradient.dart';
 import 'package:brainworld/controllers/post_controller.dart';
-import 'package:brainworld/models/user.dart';
-import 'package:brainworld/pages/chats/models/cart_model.dart';
-import 'package:brainworld/pages/chats/models/posts_model.dart';
+import 'package:brainworld/models/user_model.dart';
 import 'package:brainworld/pages/upload/course/course_desc.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
