+++ lib/controllers/chat_users_controller.dart
@@ -1,4 +1,4 @@
-import 'package:brainworld/models/user.dart';
+import 'package:brainworld/models/user_model.dart';
 import 'package:get/get.dart';
 
 class ChatUsersController extends GetxController {
