+++ lib/models/chat_model.dart
