+++ lib/models/message_model.dart
