+++ lib/models/models.dart
@@ -1,0 +1,8 @@
+export 'books_model.dart';
+export 'cart_model.dart';
+export 'chat_model.dart';
+export 'isnewuser_data_model.dart';
+export 'message_model.dart';
+export 'order_model.dart';
+export 'posts_model.dart';
+export 'user_model.dart';
