+++ lib/pages/upload/post_detail.dart
@@ -2,7 +2,7 @@ import 'package:brainworld/components/comment_fieldbox.dart';
 import 'package:brainworld/components/course_tile.dart';
 import 'package:brainworld/components/profile_user_widget.dart';
 import 'package:brainworld/constants/constant.dart';
-import 'package:brainworld/models/user.dart';
+import 'package:brainworld/models/user_model.dart';
 import 'package:brainworld/pages/chats/models/posts_model.dart';
 import 'package:brainworld/themes/themes.dart';
 import 'package:flutter/material.dart';
